requirements.txt
anyio==4.6.0
APScheduler==3.10.4
certifi==2024.8.30
charset-normalizer==3.3.2
h11==0.14.0
heroku3==5.2.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
pyaes==1.6.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
Telethon==1.37.0
tzlocal==5.2
urllib3==2.2.3
